# specs/001-frontend-dashboard/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: WeChat News Fetch Dashboard API
  version: 0.1.0
servers:
  - url: http://localhost:8000
paths:
...
  /api/accounts/{accountId}:
    patch:
      summary: Update account selection or name
      parameters:
        - in: path
          name: accountId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Account updated
    delete:
      summary: Delete account
      parameters:
        - in: path
          name: accountId
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Account deleted
...
  /api/articles/{articleId}/markdown:
    get:
      summary: Download single article as markdown
      parameters:
        - in: path
          name: articleId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Markdown file download
...